feat: implement event handling for single and multi events with notification integration
