+++ src/app/notification/service/handle-rule-notification.ts
@@ -1,0 +1,1 @@
+export const handleRuleNotification = async (): Promise<void> => {};
