+++ src/app/rule/service/handle-event/handle-multi-event.ts
@@ -1,0 +1,69 @@
+import DeventCache from '@app/devent/cache/devent-cache';
+import { Devent } from '@app/devent/entity/devent';
+import { Device } from '@app/device/entity/device';
+import { DeviceRuleAlertToLaunch } from '@app/device/entity/device-rule-alert-to-launch';
+import { RuleResultEventComparation } from '@app/rule/dto/rule-result-event-comparation';
+import { Rule } from '@app/rule/entity/rule';
+import { container } from 'tsyringe';
+import z, { object } from 'zod/v4';
+import { processEventSelector } from './process-event-selector';
+import { loggerDebug } from '@maur025/core-logger';
+import { handleRuleNotification } from '@app/notification/service/handle-rule-notification';
+
+const loggerAuxMessage: string = `[RULE] (handleMultiEvent)`;
+
+const HandleMultiEventRequest = object({
+	rule: Rule,
+	device: Device,
+});
+
+type HandleMultiEventRequest = z.infer<typeof HandleMultiEventRequest>;
+
+export const handleMultiEvent = async (
+	request: HandleMultiEventRequest,
+): Promise<DeviceRuleAlertToLaunch[]> => {
+	const { rule, device } = HandleMultiEventRequest.parse(request);
+
+	const deventCache = container.resolve(DeventCache);
+
+	const resultAndEvents: RuleResultEventComparation[] = [];
+	const resultOrEvents: RuleResultEventComparation[] = [];
+
+	for (const event of rule.events) {
+		const devent: Devent | undefined = deventCache.getById(event.deventId);
+
+		if (!devent) {
+			continue;
+		}
+
+		if (devent.condition === 'AND') {
+			resultAndEvents.push(
+				await processEventSelector({ rule, devent, device }),
+			);
+
+			continue;
+		}
+
+		resultOrEvents.push(await processEventSelector({ rule, devent, device }));
+	}
+
+	const resultOfComparation: boolean =
+		resultAndEvents.every(value => value.wasTriggered) &&
+		resultOrEvents.some(value => value.wasTriggered);
+
+	if (!resultOfComparation) {
+		loggerDebug(`${loggerAuxMessage} rule not triggered.`);
+		return [];
+	}
+
+	const andAlertLaunchList: DeviceRuleAlertToLaunch[] = resultAndEvents.flatMap(
+		andEvent => andEvent.alertToLaunchList,
+	);
+
+	const orAlertLaunchList: DeviceRuleAlertToLaunch[] = resultOrEvents.flatMap(
+		orEvent => orEvent.alertToLaunchList,
+	);
+
+	await handleRuleNotification();
+	return [...andAlertLaunchList, ...orAlertLaunchList];
+};
