+++ src/app/rule/service/handle-event/handle-single-event.ts
@@ -1,0 +1,52 @@
+import DeventCache from '@app/devent/cache/devent-cache';
+import { Devent } from '@app/devent/entity/devent';
+import { DeviceRuleAlertToLaunch } from '@app/device/entity/device-rule-alert-to-launch';
+import { RuleResultEventComparation } from '@app/rule/dto/rule-result-event-comparation';
+import { container } from 'tsyringe';
+import { processEventSelector } from './process-event-selector';
+import { loggerDebug } from '@maur025/core-logger';
+import z, { object } from 'zod/v4';
+import { Rule } from '@app/rule/entity/rule';
+import { Device } from '@app/device/entity/device';
+import { handleRuleNotification } from '@app/notification/service/handle-rule-notification';
+
+const loggerAuxMessage: string = `[RULE] (handleSingleEvent)`;
+
+const HandleSingleEventRequest = object({
+	rule: Rule,
+	device: Device,
+});
+
+type HandleSingleEventRequest = z.infer<typeof HandleSingleEventRequest>;
+
+export const handleSingleEvent = async (
+	request: HandleSingleEventRequest,
+): Promise<DeviceRuleAlertToLaunch[]> => {
+	const { rule, device } = HandleSingleEventRequest.parse(request);
+
+	const deventCache = container.resolve(DeventCache);
+
+	const devent: Devent | undefined = deventCache.getById(
+		rule.events[0].deventId,
+	);
+
+	if (!devent) {
+		return [];
+	}
+
+	const resultOfComparation: RuleResultEventComparation =
+		await processEventSelector({
+			rule,
+			devent,
+			device,
+		});
+
+	if (!resultOfComparation.wasTriggered) {
+		loggerDebug(`${loggerAuxMessage} rule not triggered.`);
+
+		return [];
+	}
+
+	await handleRuleNotification();
+	return resultOfComparation.alertToLaunchList;
+};
